project/requirements.txt
Django==2.2.6
django-bootstrap-form==3.4
django-environ==0.4.5
django-excel==0.0.10
django-filter==2.2.0
entrypoints==0.3
environ==1.0
et-xmlfile==1.0.1
flake8==3.7.8
jdcal==1.4.1
lml==0.0.9
mccabe==0.6.1
openpyxl==2.5.14
pep8==1.7.1
psycopg2>=2.7,<3.0
pycodestyle==2.5.0
pyexcel==0.5.15
pyexcel-io==0.5.20
pyexcel-webio==0.1.4
pyexcel-xls==0.5.8
pyexcel-xlsx==0.5.7
pyflakes==2.1.1
pytz==2019.3
sqlparse==0.3.0
texttable==1.6.2
xlrd==1.2.0
xlwt==1.3.0
